GUI: UIContext: move FocusCycle to ui/focus.rs; no behavior changes
